Move alias logic into separate file
