+++ webpack.config.base.js
@@ -3,32 +3,12 @@
 const path = require('path');
 const webpack = require('webpack');
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
+const generateReactAlias = require('./webpack/generate-alias');
 
 // React doesn't like being included multiple times as can happen when using
 // yarn link. Here we find a more specific path to it by first looking in
 // stripes-core (__dirname) before falling back to the platform or simply react
-let specificReact;
-const platformReact = path.join(path.resolve(), 'node_modules', 'react');
-const coreReact = path.join(__dirname, 'node_modules', 'react');
-const fallbackReact = 'react';
-
-if ( tryResolve(coreReact)) {
-  specificReact = coreReact;
-} else if ( tryResolve(platformReact)) {
-  specificReact = platformReact;
-} else {
-  specificReact = fallbackReact;
-}
-
-function tryResolve(path){
-  try {
-    require.resolve(path);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
-
+let specificReact = generateReactAlias();
 
 module.exports = {
   entry: [
